test: 시간 문자열을 int로 변환하는 유틸 클래스 메소드 테스트 코드 작성
